Add setBatch method on the properyList object. Using json2.js if the browser does not support JSON.stringify.
git-svn-id: https://svn.open.seria.net/svn/platform/trunk/seria@3864 d23b63ab-52c4-41c5-9713-938aad2d84f3

+++ components/SERIA_JsonUserData/api/setProperty.php
@@ -19,10 +19,18 @@ try {
 
 	$storage = new SERIA_JsonUserPropertyStorageDriver($user);
 
-	if (isset($_REQUEST['value']))
-		$storage->set($_REQUEST['namespace'], $_REQUEST['name'], $_REQUEST['value']);
-	else
-		$storage->delete($_REQUEST['namespace'], $_REQUEST['name']);
+	if (isset($_REQUEST['name'])) {
+		if (isset($_REQUEST['value']))
+			$storage->set($_REQUEST['namespace'], $_REQUEST['name'], $_REQUEST['value']);
+		else
+			$storage->delete($_REQUEST['namespace'], $_REQUEST['name']);
+	} else if (isset($_REQUEST['batch'])) {
+		$batch = SERIA_Lib::fromJSON($_REQUEST['batch']);
+		foreach ($batch as $name => $value) {
+			$storage->set($_REQUEST['namespace'], $name, $value);
+		}
+	} else
+		throw new SERIA_Exception('setProperty: Invalid argument!');
 
 	SERIA_Lib::publishJSON(array(
 		'error' => false
